// client/src/model/options/apartment-options.ts
export class ApartmentOptions {
  water: string[] = [];
  gas: string[] = [];
  electricity: string[] = [];
  sewer: string[] = [];
  condition: string[] = [];
  walls: string[] = [];
  hotWater: string[] = [];
  heating: string[] = [];
  bath: string[] = [];
  toilet: string[] = [];
  constructionTime: string[] = [];
  loggia: string[] = [];
  balcony: string[] = [];
}